viperHTML meets csso and uglify-js minifier - - - - - - - - - - - - - - - - - - - - - -
Whenever a tag is a style or a script one, and its content is static,
there is no reason to not minify it once per template literal.

Both css and js minifiers are also exported as helper so that,
whenever it's needed, users can minify on the fly too.

viperHTML.minify.css('text');
viperHTML.minify.js('text');

Both methods accept optional options argument.