Mark comments around interpolations and with unique IDs
The previous update was flagging the end of each interpolation
but it was still ambiguous/hard to figure out on the client side
how to adopt such content.

This version uses similar approach but it marks through unique content
begin and end of each interpolation.

The goal is to simplify hyperHTML adoption and make it infallible
with content created through viperHTML.